+++ cmd/nex/helpers_windows.go
@@ -1,0 +1,8 @@
+package main
+
+func checkVer(globals *Globals) error {
+	if globals.Check {
+		return nil
+	}
+	return nil
+}
